Trying to move compositor network creation to ensure_properties()
